Add docs for Modal
